Refactor default values and make it work in loop control
