+++ src/mixins/index.js
@@ -3,3 +3,4 @@ export { default as shouldElementBeVisible } from './shouldElementBeVisible';
 export { default as getValidPath } from './getValidPath';
 export { default as formWatchers } from './formWatchers';
 export { default as multiselectApi } from './multiselectApi';
+export { default as defaultValues } from './defaultValues';
